Extract models & interfaces to dedicated project
