+++ RevitTranslator.Abstractions/Models/Views/ViewTypeInternal.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel;
 
-namespace RevitTranslator.Revit.Core.Models;
+namespace RevitTranslator.Abstractions.Models.Views;
 
 public enum ViewTypeInternal
 {
