--- RevitTranslator.Revit.Abstractions/RevitTranslator.Revit.Abstractions.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <OutputType>Library</OutputType>
-        <TargetFrameworks>net48;net8.0</TargetFrameworks>
-        <RepositoryUrl>https://github.com/Krachkovskii/RevitTranslator</RepositoryUrl>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="PolySharp" />
-    </ItemGroup>
-
-</Project>
