+++ RevitTranslator.Revit.Core.Abstractions/Contracts/IRevitContextProvider.cs
@@ -1,0 +1,8 @@
+using Autodesk.Revit.DB;
+
+namespace RevitTranslator.Revit.Abstractions.Contracts;
+
+public interface IRevitContextProvider
+{
+    Document ActiveDocument { get; }
+}
