+++ RevitTranslator.Revit.Core.Abstractions/Contracts/IRevitHandler.cs
@@ -1,0 +1,6 @@
+namespace RevitTranslator.Revit.Abstractions.Contracts;
+
+public interface IRevitHandler
+{
+    Task RaiseAsync(Action handler);
+}
