+++ RevitTranslator.Revit.Core.Abstractions/RevitTranslator.Revit.Core.Abstractions.csproj
@@ -1,0 +1,34 @@
+<Project Sdk="Microsoft.NET.Sdk">
+    <PropertyGroup>
+        <OutputType>Library</OutputType>
+        <Configurations>Debug R23;Debug R24;Debug R25;Debug R26</Configurations>
+        <Configurations>$(Configurations);Release R23;Release R24;Release R25;Release R26</Configurations>
+        <RepositoryUrl>https://github.com/Krachkovskii/RevitTranslator</RepositoryUrl>
+        <RootNamespace>RevitTranslator.Revit.Abstractions</RootNamespace>
+    </PropertyGroup>
+
+    <PropertyGroup Condition="$(Configuration.Contains('R23'))">
+        <RevitVersion>2023</RevitVersion>
+        <TargetFramework>net48</TargetFramework>
+    </PropertyGroup>
+    <PropertyGroup Condition="$(Configuration.Contains('R24'))">
+        <RevitVersion>2024</RevitVersion>
+        <TargetFramework>net48</TargetFramework>
+    </PropertyGroup>
+    <PropertyGroup Condition="$(Configuration.Contains('R25'))">
+        <RevitVersion>2025</RevitVersion>
+        <TargetFramework>net8.0</TargetFramework>
+        <GenerateDependencyFile>false</GenerateDependencyFile>
+    </PropertyGroup>
+    <PropertyGroup Condition="$(Configuration.Contains('R26'))">
+        <RevitVersion>2026</RevitVersion>
+        <TargetFramework>net8.0</TargetFramework>
+        <GenerateDependencyFile>false</GenerateDependencyFile>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="PolySharp"/>
+        <PackageReference Include="Nice3point.Revit.Api.RevitAPI"/>
+    </ItemGroup>
+    
+</Project>
