+++ RevitTranslator.Revit.Core/Services/TranslationReportService.cs
@@ -2,6 +2,8 @@ using System.Diagnostics;
 using System.Globalization;
 using System.IO;
 using System.Text;
+using RevitTranslator.Abstractions;
+using RevitTranslator.Abstractions.Contracts;
 using RevitTranslator.Revit.Core.Contracts;
 using RevitTranslator.Revit.Core.Extensions;
 using RevitTranslator.Revit.Core.Models;
