+++ RevitTranslator.UI.Demo/Utils/MockTranslationReportService.cs
@@ -1,4 +1,4 @@
-using RevitTranslator.Revit.Core.Contracts;
+using RevitTranslator.Abstractions.Contracts;
 
 namespace RevitTranslator.UI.Demo.Utils;
