+++ RevitTranslator.UI/FallbackValues/CategoryTypeFallbackDescriptor.cs
@@ -1,7 +1,6 @@
 using System.Collections.ObjectModel;
+using RevitTranslator.Abstractions.Models.Categories;
 using RevitTranslator.Common.Extensions;
-using RevitTranslator.Common.Models;
-using RevitTranslator.Common.Models.Categories;
 
 namespace RevitTranslator.UI.FallbackValues;
