+++ RevitTranslator.UI/Messages/ViewsSelectedMessage.cs
@@ -1,4 +1,5 @@
-using RevitTranslator.Revit.Core.Models;
+using RevitTranslator.Abstractions.Models;
+using RevitTranslator.Abstractions.Models.Views;
 
 namespace RevitTranslator.UI.Messages;
