+++ RevitTranslator.UI/Models/ViewGroupViewModel.cs
@@ -1,7 +1,8 @@
 using System.Collections.ObjectModel;
 using System.ComponentModel;
 using CommunityToolkit.Mvvm.ComponentModel;
-using RevitTranslator.Revit.Core.Models;
+using RevitTranslator.Abstractions.Models;
+using RevitTranslator.Abstractions.Models.Views;
 
 namespace RevitTranslator.UI.Models;
