+++ RevitTranslator.UI/ViewModels/SettingsViewModel.cs
@@ -5,7 +5,8 @@ using System.Threading.Tasks;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using CommunityToolkit.Mvvm.Messaging;
-using RevitTranslator.Revit.Core.Contracts;
+using RevitTranslator.Abstractions;
+using RevitTranslator.Abstractions.Contracts;
 using RevitTranslator.Common.Messages;
 using TranslationService.Models;
 using TranslationService.Utils;
